requirements.txt
aiolimiter==1.1.0
anyio==3.6.2
APScheduler==3.10.4
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
bcrypt==4.0.1
black==23.1.0
bottle==0.12.25
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.2
Django==4.0.8
docopt==0.6.2
exceptiongroup==1.1.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
iniconfig==2.0.0
markdown-it-py==2.2.0
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
paramiko==3.0.0
pathspec==0.11.0
pipreqs==0.4.11
platformdirs==3.1.0
pluggy==1.3.0
psutil==5.9.4
psycopg2-binary==2.9.5
pycparser==2.21
Pygments==2.14.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pytest==7.4.0
python-dateutil==2.8.2
python-decouple==3.8
python-telegram-bot==20.4
pytz==2022.7.1
redis==4.5.3
reloading==1.1.2
requests==2.28.2
rfc3986==1.5.0
rich==13.3.3
rich_argparse==1.1.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
sqlparse==0.4.3
sshtunnel==0.4.0
starlette==0.25.0
telegram==0.0.1
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==1.26.14
uvicorn==0.20.0
yarg==0.1.9
